+++ src/tar.c
@@ -2224,8 +2224,8 @@ parse_opt (int key, char *arg, struct argp_state *state)
 
     case ARGP_KEY_ERROR:
       if (args->loc->source == OPTS_FILE)
-	error (0, 0, _("%s:%lu: location of the error"), args->loc->name,
-	       (unsigned long) args->loc->line);
+	error (0, 0, _("%s:%jd: location of the error"), args->loc->name,
+	       args->loc->line);
       else if (args->loc->source == OPTS_ENVIRON)
 	error (0, 0, _("error parsing %s"), args->loc->name);
       exit (EX_USAGE);
